requirements.txt
torchvision==0.9.0
torch==1.8.1
Pillow==8.1.2
PyYAML==5.4.1
